webfaf2: rename webfaf2 module to webfaf2_main
Renamed to not conflict with the webfaf2 package.
