+++ src/webfaf2/hub.wsgi
@@ -7,4 +7,4 @@ logging.basicConfig(stream=sys.stderr)
 os.environ["WEBFAF_ENVIRON_PRODUCTION"] = "1"
 sys.path.insert(0, os.path.dirname(__file__))
 
-from webfaf2 import app as application
+from webfaf2_main import app as application
