+++ src/webfaf2/reports.py
@@ -48,7 +48,7 @@ from utils import (Pagination,
 
 reports = Blueprint("reports", __name__)
 
-from webfaf2 import db
+from webfaf2_main import db
 from forms import (ReportFilterForm, NewReportForm, NewAttachmentForm,
                    component_list)
